--- user_service/user/src/user_service_api/mod.rs
@@ -1,2 +1,0 @@
-pub mod models;
-pub mod user_service;
