--- user_service/user/src/user_service_api/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod user;
-pub mod user_registration;
-pub mod user_login;
