--- user_service/user/src/user_service_api/models/user_login.rs
@@ -1,6 +1,0 @@
-/// UserLogin is used to map the details at ClientRequest time for login end-point
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
-pub struct UserLogin {
-    pub email: String,
-    pub password: String,
-}
