--- user_service/user/src/user_service_api/models/user_registration.rs
@@ -1,7 +1,0 @@
-/// UserRegistration is used to map the details at ClientRequest time at create_user end-point
-#[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
-pub struct UserRegistration {
-    pub name: String,
-    pub email: String,
-    pub password: String,
-}
