--- user_service/user/src/user_service_api/user_service/mod.rs
@@ -1,1 +1,0 @@
-pub mod handler;
