--- user_service/user/src/user_service_impl/constants/constant.rs
@@ -1,9 +1,0 @@
-pub static DATABASE_PORT_ADDRESS: &str = "127.0.0.1:9042";
-
-pub static SERVER_BIND_PORT: &str = "127.0.0.1:3080";
-
-pub static INDEX: usize = 0;
-
-pub static DEBUG_LEVEL_KEY: &str = "RUST_LOG";
-
-pub static DEBUG_LEVEL_VALUE: &str = "actix_web=debug";
