--- user_service/user/src/user_service_impl/constants/mod.rs
@@ -1,2 +1,0 @@
-pub mod constant;
-pub mod queries;
