--- user_service/user/src/user_service_impl/controller/mod.rs
@@ -1,1 +1,0 @@
-pub mod error;
