--- user_service/user/src/user_service_impl/env_setup/mod.rs
@@ -1,4 +1,0 @@
-pub mod connection;
-pub mod keyspace;
-pub mod table;
-pub mod set_up;
