--- user_service/user/src/user_service_impl/eventsourcing/user_command/mod.rs
@@ -1,1 +1,0 @@
-pub mod models;
