+++ user_service/user/src/user_service_impl/eventsourcing/user_event.rs
@@ -1,4 +1,4 @@
-use crate::user_service_impl::models::p_user::PUser;
+use crate::user_service_impl::eventsourcing::user_entity::PUser;
 
 /// UserEvent is a struct which holds all events related to User
 #[derive(Serialize, Deserialize, Debug, Clone, Event)]
