--- user_service/user/src/user_service_impl/eventsourcing/user_repository/mod.rs
@@ -1,3 +1,0 @@
-pub mod insertion;
-pub mod display;
-pub mod is_present;
