+++ user_service/user/src/user_service_impl/eventsourcing/user_state.rs
@@ -1,5 +1,5 @@
-use crate::user_service_impl::models::p_user::PUser;
 use eventsourcing::{prelude::*};
+use crate::user_service_impl::eventsourcing::user_entity::PUser;
 
 ///UserState is a struct which holds persistent user and its generation
 #[derive(Debug, Deserialize, Serialize, Clone, PartialEq)]
@@ -9,8 +9,9 @@ pub struct UserState {
 }
 
 impl AggregateState for UserState {
-    #[cfg_attr(tarpaulin,skip)]
+    #[cfg_attr(tarpaulin, skip)]
     fn generation(&self) -> u64 {
         self.generation
     }
 }
+
