+++ user_service/user/src/user_service_impl/mod.rs
@@ -1,11 +1,5 @@
 pub mod env_setup;
 
-pub mod models;
-
 pub mod eventsourcing;
 
-pub mod controller;
-
-pub mod constants;
-
-pub mod utilities;
+pub mod handler;
