--- user_service/user/src/user_service_impl/models/mod.rs
@@ -1,2 +1,0 @@
-pub mod p_user;
-pub mod get_user;
