--- user_service/user/src/user_service_impl/utilities/mod.rs
@@ -1,3 +1,0 @@
-pub mod mappers;
-pub mod initial_state;
-pub mod wrapper;
